server/requirements.txt
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
praat-parselmouth>=0.4.4
numpy>=1.24.0
